requirements.txt
accelerate>=1.2.0
datasets>=3.2.0
git+https://github.com/huggingface/diffusers.git
git+https://github.com/huggingface/transformers.git
huggingface_hub>=0.34.0
Pillow>=10.4.0
PyYAML>=6.0.2
safetensors>=0.5.0
tqdm>=4.67.1
